chore: rename `shad-lib` to `lib`, also updated test snapshots for adding type to `add button` and `icon button`
