+++ packages/shadcn/components.json
@@ -14,7 +14,7 @@
     "components": "src/components",
     "utils": "src/lib/utils",
     "ui": "src/components/ui",
-    "lib": "src/shad-lib",
+    "lib": "src/lib",
     "hooks": "src/hooks"
   },
   "iconLibrary": "lucide"
