+++ packages/shadcn/src/FieldHelpTemplate/FieldHelpTemplate.tsx
@@ -1,5 +1,5 @@
 import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId } from '@rjsf/utils';
-import { cn } from '../shad-lib/utils';
+import { cn } from '../lib/utils';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
