+++ packages/shadcn/src/components/ui/command.tsx
@@ -5,7 +5,7 @@ import { type DialogProps } from '@radix-ui/react-dialog';
 import { Command as CommandPrimitive } from 'cmdk';
 import { Search } from 'lucide-react';
 
-import { cn } from '../../shad-lib/utils';
+import { cn } from '../../lib/utils';
 import { Dialog, DialogContent } from './dialog';
 
 const Command = React.forwardRef<
