+++ packages/shadcn/src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '../../shad-lib/utils';
+import { cn } from '../../lib/utils';
 
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
